+++ src/main/java/com/moodmate/moodmatebe/domain/chat/domain/Message.java
@@ -1,0 +1,2 @@
+package com.moodmate.moodmatebe.domain.chat.domain;public class Message {
+}
