--- src/main/java/com/moodmate/moodmatebe/domain/chat/dto/ChatPageableDto.java
@@ -1,13 +1,0 @@
-package com.moodmate.moodmatebe.domain.chat.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ChatPageableDto {
-    private int size;
-    private int page;
-    private int totalPages;
-    private int totalElements;
-}
