--- src/main/java/com/moodmate/moodmatebe/domain/chat/dto/RedisChatMessageDto.java
@@ -1,38 +1,0 @@
-package com.moodmate.moodmatebe.domain.chat.dto;
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
-import lombok.*;
-
-import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
-
-@Getter
-@Builder
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class RedisChatMessageDto {
-    private Long messageId;
-    private Long userId;
-    private Long roomId;
-    private String content;
-    private Boolean isRead;
-    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-    @JsonSerialize(using = LocalDateTimeSerializer.class)
-    private LocalDateTime createdAt;
-
-    public Map<String, Object> toMap() {
-        Map<String, Object> map = new HashMap<>();
-        map.put("messageId", messageId);
-        map.put("userId", userId);
-        map.put("roomId", roomId);
-        map.put("content", content);
-        map.put("isRead", isRead);
-        map.put("createdAt", createdAt);
-        return map;
-    }
-}
