+++ src/main/java/com/moodmate/moodmatebe/domain/chat/dto/request/ChatMessageDto.java
@@ -2,14 +2,20 @@ package com.moodmate.moodmatebe.domain.chat.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.ToString;
+
+import java.time.LocalDateTime;
 
 @Getter
 @ToString
+@NoArgsConstructor
 @AllArgsConstructor
 public class ChatMessageDto {
     private String content;
     private Long roomId;
     private String token;
     private Long userId;
+    private LocalDateTime createdAt;
+
 }
