+++ src/main/java/com/moodmate/moodmatebe/domain/chat/repository/MessageRepository.java
@@ -1,18 +1,12 @@
 package com.moodmate.moodmatebe.domain.chat.repository;
 
-import com.moodmate.moodmatebe.domain.chat.domain.ChatMessage;
-import com.moodmate.moodmatebe.domain.chat.domain.ChatRoom;
+import com.moodmate.moodmatebe.domain.chat.domain.Message;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
+import org.springframework.data.mongodb.repository.MongoRepository;
 
-import java.util.List;
+public interface ChatMessageRepository extends MongoRepository<Message, Long> {
+    Page<Message> findByRoomIdOrderByCreatedAtDesc(Long roomId, Pageable pageable);
 
-public interface MessageRepository extends JpaRepository<ChatMessage, Long> {
-
-    @Query("SELECT ch FROM chat_message ch WHERE ch.room = :room ORDER BY ch.createdAt DESC")
-    List<ChatMessage> findByRoomOrderByCreatedAt(@Param("room") ChatRoom room, Pageable pageable);
-
-    int countByRoom(ChatRoom room);
+    int countByRoomId(Long roomId);
 }
