+++ src/main/java/com/moodmate/moodmatebe/global/config/MongoDBConfig.java
@@ -1,0 +1,2 @@
+package com.moodmate.moodmatebe.global.config;public class MongoDBConfig {
+}
