+++ src/main/resources/static/index.html
@@ -1,0 +1,126 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>WebSocket Chat</title>
+    <script src="https://code.jquery.com/jquery-3.6.4.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/sockjs-client@1/dist/sockjs.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/stomp.js/2.3.3/stomp.min.js"></script>
+</head>
+
+<body>
+<h1>WebSocket Chat</h1>
+
+<div>
+    <label for="user">Username:</label>
+    <input type="text" id="user" placeholder="Enter your username">
+</div>
+
+<div>
+    <label for="roomIdInput">Room ID:</label>
+    <input type="text" id="roomIdInput" placeholder="Enter Room ID">
+</div>
+
+<div>
+    <label for="userIdInput">User ID:</label>
+     <input type="text" id="userIdInput" placeholder="Enter User ID">
+</div>
+
+<div id="chatArea" style="border: 1px solid #ccc; padding: 10px; margin-bottom: 10px; max-height: 200px; overflow-y: auto;"></div>
+
+<input type="text" id="messageInput" placeholder="Enter your message">
+<button onclick="connect()">Connect</button>
+<button onclick="sendMessage()">Send Message</button>
+<button onclick="disconnect()">Disconnect</button>
+
+<script>
+    var stompClient = null;
+    var username = null;
+    var roomId = null;
+    var userId = null;
+
+    function connect() {
+        username = document.getElementById('user').value;
+        roomId = document.getElementById('roomIdInput').value;
+        userId = document.getElementById('userIdInput').value;
+
+        var socket = new SockJS('/chat');
+        stompClient = Stomp.over(socket);
+
+        stompClient.connect({}, function onSuccess(frame) {
+            console.log('Connected: ' + frame);
+
+            stompClient.send('/pub/chat', {}, JSON.stringify({
+                sender: username,
+                content: 'has joined the chat',
+                roomId: roomId,
+                userId: userId,
+            }));
+
+            stompClient.subscribe('/sub/chat/' + roomId, function (response) {
+                var message = JSON.parse(response.body);
+                console.log('Parsed message:', message);
+
+                displayMessage(message.userNo, message.content);
+            });
+
+            sessionStorage.setItem('username', username);
+            sessionStorage.setItem('roomId', roomId);
+            sessionStorage.setItem('userId', userId);
+
+            syncState();
+        }, function onError(error) {
+            console.log('Connection error: ' + error);
+        });
+    }
+
+    function syncState() {
+        var storedUsername = sessionStorage.getItem('username');
+        var storedRoomId = sessionStorage.getItem('roomId');
+        var storedUserId = sessionStorage.getItem('userId');
+
+        if (storedUsername && storedRoomId == roomId && storedUserId == userId) {
+            username = storedUsername;
+            document.getElementById('user').value = username;
+            document.getElementById('roomIdInput').value = storedRoomId;
+            document.getElementById('userIdInput').value = storedUserId;
+        }
+    }
+
+    function sendMessage() {
+        if (stompClient && stompClient.connected) {
+            var messageContent = document.getElementById('messageInput').value;
+
+            stompClient.send('/pub/chat', {}, JSON.stringify({
+                userId: userId,
+                roomId: roomId,
+                content: messageContent,
+            }));
+        }
+    }
+
+    function disconnect() {
+        stompClient.send('/pub/chat', {}, JSON.stringify({
+            sender: username,
+            content: 'has left the chat',
+            roomId: roomId,
+            userNo: userId
+        }));
+
+        stompClient.disconnect();
+    }
+
+    function displayMessage(sender, content) {
+        var chatArea = document.getElementById('chatArea');
+        var messageElement = document.createElement('div');
+        messageElement.innerHTML = '<strong>' + sender + ':</strong> ' + content;
+        chatArea.appendChild(messageElement);
+
+        chatArea.scrollTop = chatArea.scrollHeight;
+    }
+</script>
+</body>
+
+</html>
